+++ src/main/java/com/putterfly/instructions/factories/BreakFactory.java
@@ -5,7 +5,7 @@ import com.putterfly.instructions.cat1.Break;
 
 public class BreakFactory implements CommandFactory{
     @Override
-    public Command newCommand(long instruction) {
+    public Command newCommand(int instruction) {
         return new Break(instruction);
     }
 }
