+++ src/main/java/com/putterfly/instructions/factories/CommandFactory.java
@@ -3,5 +3,5 @@ package com.putterfly.instructions.factories;
 import com.putterfly.instructions.Command;
 
 public interface CommandFactory {
-    public Command newCommand(long instruction);
+    public Command newCommand(int instruction);
 }
