+++ src/main/java/com/putterfly/simulator/ProgramCounter.java
@@ -1,17 +1,17 @@
 package com.putterfly.simulator;
 
 public class ProgramCounter {
-    private static long pc;
+    private static int pc;
 
-    public static long getPC(){
+    public static int getPC(){
         return pc;
     }
 
-    public static void setPC(long address){
+    public static void setPC(int address){
         pc = address;
     }
 
-    public static void advancePC(long offset){
+    public static void advancePC(int offset){
         pc += offset;
     }
 }
