+++ src/main/java/com/putterfly/simulator/Register.java
@@ -1,8 +1,8 @@
 package com.putterfly.simulator;
 
 public class Register {
-    private static long[] registers;
-    public static long getRegisterValue(int index){
+    private static int[] registers;
+    public static int getRegisterValue(int index){
         return registers[index];
     }
     public static void setRegisterValue(int index,int value){
